refactor: remove the drill-down0layout
