+++ packages/s2-core/src/utils/index.ts
@@ -1,7 +1,7 @@
 export { splitTotal } from './data-set-operate';
 export { auto } from './formatter';
 export { measureTextWidth } from './text';
-export * from './drill-down/helper';
+export * from './drill-down';
 export * from './export/index';
 export * from './interaction/merge-cells';
 export * from './g-renders';
